+++ qiskit/ignis/verification/randomized_benchmarking/__init__.py
@@ -1,18 +1,26 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2019, IBM.
+# This code is part of Qiskit.
 #
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
 
 
 """
-Measurement correction module
+Randomized Benchmarking module
 """
 
-# Measurement correction functions
+# Randomized Benchmarking functions
 from .Clifford import Clifford
-from . import clifford_utils
+from .basic_utils import BasicUtils
+from .clifford_utils import CliffordUtils
 from .circuits import randomized_benchmarking_seq
-from .fitters import RBFitter
+from .fitters import RBFitter, InterleavedRBFitter, PurityRBFitter
 from . import rb_utils