+++ qiskit/ignis/verification/tomography/basis/paulibasis.py
@@ -1,9 +1,16 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2018, IBM.
+# This code is part of Qiskit.
 #
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
 
 """
 Pauli tomography preparation and measurement basis
@@ -34,7 +41,7 @@ def pauli_measurement_circuit(op, qubit, clbit):
         A QuantumCircuit object.
     """
 
-    circ = QuantumCircuit(qubit[0], clbit[0])
+    circ = QuantumCircuit(qubit.register, clbit.register)
     if op == 'X':
         circ.h(qubit)
         circ.measure(qubit, clbit)
@@ -61,7 +68,7 @@ def pauli_preparation_circuit(op, qubit):
         A QuantumCircuit object.
     """
 
-    circ = QuantumCircuit(qubit[0])
+    circ = QuantumCircuit(qubit.register)
     if op == 'Xp':
         circ.h(qubit)
     if op == 'Xm':