+++ test/tomography/test_state_tomography.py
@@ -1,9 +1,16 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2019, IBM.
+# This code is part of Qiskit.
 #
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
 
 # pylint: disable=missing-docstring
 
@@ -90,6 +97,7 @@ class TestStateTomography(unittest.TestCase):
 
     def test_complex_3_qubit_circuit(self):
         def rand_angles():
+            # pylint: disable=E1101
             return tuple(2 * numpy.pi * numpy.random.random(3) - numpy.pi)
 
         q = QuantumRegister(3)