+++ CHANGELOG.md
@@ -19,12 +19,39 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Pulse Discriminator (\#238, \#278)
+
+## [0.2.0](https://github.com/Qiskit/qiskit/compare/0.1.1...0.2.0)- 2019-08-22
+
+### Added
+
+- Logging Module (\#153)
+- Purity RB (\#218)
+- Interleaved RB (\#174)
+- Repetition Code for Verification (\#210)
+
 ### Changed
+
+- Apply measurement mitigation in parallel when applied to multiple results (\#240)
+- Add multiple results to measurement mitigation (\#240)
+- Fixed bug in RB fit error
+- Updates for Terra Qubit class (\#200)
+- Added the ability to add arbitrary seeds to RB (not just in order) (\#208)
+- Fix bug in the characterization fitter when selecting a qubit index to fit
+- Improved guess values for RB fitters and enabled the user to input their own guess values
 
 ### Removed
 
+## [0.1.1] - 2019-05-02
 
-## [0.1.0] - 2019-TBD
+### Added
+
+- Tensored Measurement Mitigation
+- Align cliffs option to RB
+- Quantum Volume
+- Subset measurement mitigation
+
+## [0.1.0] - 2019-03-04
 
 ### Added
 
@@ -32,4 +59,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 
 [UNRELEASED]: TBD
-[0.1.0]: TBD
+[0.1.0]: March 4, 2019